refactor: moved event-panel-view actions-buttons to a separate file to allow testing
